requirements/base.txt
django==2.2.*
django_extensions==2.1.*
djangorestframework==3.9.*
drf-nested-routers==0.91
drf_yasg==1.11.*
django-cors-headers==2.4.*

django-coverage-plugin==1.6.*
factory-boy==2.12.*
flake8==3.7.*
pycodestyle==2.5.*
pytest==3.10.1
pytest-cov==2.6.*
pytest-xdist==1.28.*
pytest-django==3.4.*
pytest-factoryboy==2.0.*
pytest-sugar==0.9.*
